requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.0
pyarrow>=14.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pandas_ta>=0.3.14b0